`resultTransformer` and `resultComparator`
You can provide a `resultTransformer` when configuring `ApolloClient` to be able to modify query results immediately before they are returned to the application.  Generally, transformers should avoid mutations of their input, as it will bust Apollo's result equality comparsion (and cause unnecessary re-renders).

For cases where it is preferable to mutate the input, you can also override how query results are compared for equality.  As an example, an application may wish to attach prototypes to result data (for read-only helpers) - it would need to modify the equality check to deep compare the two results while ignoring prototypes.  If the newly returned result (without prototypes) matches the data present in the previous result (with prototypes), Apollo would consider it a no-op and not trigger any changes.