Added an analyze script
Added a new analyze script to the package.json that makes it possible to inspect what modules make it into the published package and how efficiently it bundles in a webpack v2 project.

The script runs a compilation and builds the analyze test project (that only imports this library) with webpack and the webpack-bundle-analyzer plugin in production mode. Once complete the plugin opens  your browser to an interactive analysis at http://localhost:8888.

See discussion: https://github.com/apollostack/apollo-client/issues/684#issuecomment-266983921.