Document `setVariables` internal API status
This clarifies `setVariables` as for internal use only because it has
a severe caveat where a result coming from cache will not notify subscribers.
That makes it a poor choice for most common use cases.

Instead `refetch` should be used and is now documented as such.

Also clarifies that `watchQuery` does not watch the result, but the cache store,
meaning that if the store does not change, the subcribers will not be notified.
In practice that means that a result coming from cache will never be seen by
subscribers.

This partially address #2712