Reimplement startGraphQLSubscription using getObservableFromLink.
When outside contributors dive into the QueryManager, one of the most
common complaints is that startGraphQLSubscription adds a bunch of code
that seems to duplicate the functionality of watchQuery and fetchRequest,
which is especially unfortunate if you're not using subscriptions.

Thanks to my new getObservableFromLink helper method, we can share most of
the implementation of subscriptions with queries and mutations. With this
commit, startGraphQLSubscription is now small enough that no one should
have any reason to complain about it ever again.