Track a single Apollo Client React context via global state
This is a follow up to PR #5561. While forcing CommonJS for
all references to the Apollo context helps ensure the same React
context is used for all Apollo Client based application bundles,
it's a bit of a kludge. One off workarounds are a pain to maintain
in the long term, and are more than likely to trip us up at some
point. This approach was put in place to try to avoid using global
state to track a single React context, but after further reflection
avoiding global state here is more than likely detrimental.

This commit creates and tracks a single React context for use with
Apollo Client, that is stored on the shared React instance. This
helps ensure that only one React context is used by Apollo Client,
across all bundles.