Adjust `fetchData` to accommodate #5565 (#5639)
* Adjust `fetchData` to accommodate #5565

The changes made in #5565 mean that `QueryData#fetchData` should
no longer directly rely on `ObservableQuery#getCurrentResult` or
`ObservableQuery#result`. This commit adjusts `fetchData` such
that it returns a new Promise that resolves when the
`ObservableQuery` subscription receives new data, re-using the
same `startQuerySubscription` business rules that non-SSR renders
adhere to. This ensures that the `queryPromises` (in
`react/ssr/RenderPromises`) stored for SSR resolve properly after
new data is received.

* Remove no longer valid Promise generic