Pass Policies object via options instead of using `this`.
TypeScript doesn't do a great job of enforcing parameter types (including
`this:` types) for functions called with Function.prototype.{call,apply},
which is especially frustrating because you pretty much have to use those
methods when you want to call a function with a specific `this` object.

Another reason to avoid using `this` is simply that some developers prefer
arrow functions, and arrow functions ignore any `this` object provided by
.call or .apply. Instead, we can expose the Policies object as a property
in the FieldFunctionOptions parameter, so it can be used (or ignored)
without having to think about `this` at all.