Assume keyArgs:false when read or merge function provided.
As I've been working with the new typePolicies configuration API, I've
become increasingly convinced that the presence of a read or merge
function in a FieldPolicy should imply keyArgs:false by default, because
those functions will almost always want full control over the
interpretation of arguments.

If the developer defines a read or merge function but still wants the
cache to differentiate field values according to certain arguments, that's
easy: just add an explicit keyArgs:[...] property to the FieldPolicy, to
override the inferred keyArgs:false.