Simplify EntityStore#merge and storeObjectReconciler.
Replacing a Reference in the cache is always safe, even with
non-normalized data, since the Reference refers to data safely stored
elsewhere in the cache.

By removing this invariant, we remove the need for most of the surrounding
logic in the storeObjectReconciler function, too, which is why it can be
so much shorter now.