Delete only from layers that contain dataId in EntityStore#evict.
Eviction ignores cache layers and guarantees that the given entity ID
and/or field name will be completely removed from all layers of the cache.
As an optimization, we can avoid calling this.delete for any layers that
do not contain a this.data[dataId] entry for the given dataId.