Provide Policies to EntityStore.
I have been reluctant to give the EntityStore any knowledge of the
Policies class, but the time has come.

Both of these classes are implementation details of InMemoryCache, so
giving the EntityStore access to Policies should have no consequences for
application code. As long as the EntityStore treats store.policies as a
declarative source of policy information, their relationship should only
make it easier to enforce those policies consistently and efficiently.