Allow keyArgs functions to return KeySpecifier arrays. (#5913)
Sometimes the key arguments for a field need to be determined dynamically,
but that shouldn't mean the keyArgs function has to perform its own stable
serialization. Allowing keyArgs functions to return a KeySpecifier array
makes it easy to report which fields matter, and in what order they should
be serialized.

Small breaking change: keyArgs functions now receive the arguments object
as their first parameter, rather than the FieldNode. Other details
(including the field) may be found in the context parameter.