Mutate update function DataProxy/ApolloCache changes
This commit gives the mutate `update` function direct
access to the cache, instead of using a scaled down
`DataProxy`. While the `DataProxy` approach helped in the
past as a way to batch cache updates, it is no longer
necessary. We are only ever passing a cache instance into
`update`, which means limiting the type to be a `DataProxy`
gives us no additional advantages. The only other `DataProxy`
implementor in the AC codebase is `ApolloClient`, which
is never passed to the `update` function.

This commit also exposes optional `identify` and `modify`
API elements in `ApolloCache`. Since `update` now has
access to the full `ApolloCache` instance passed in,
`identify` and `modify` can be used in the update function.
Note that the `ApolloCache` implementations of `identify`
and `modify` do nothing of value; they should be overridden
by `ApolloCache` subclasses that are interested in using them,
like `InMemoryCache` does. By including them in `ApolloCache`
as basically no-op methods, we're avoiding forcing
alternative cache implementations that subclass `ApolloCache`
to have to support them.