Make QueryManager#getQuery register new QueryInfo.
One gotcha when using getQuery was that it might return a disposable
QueryInfo object, without storing it anywhere, without warning.

The setQuery method does not have this problem, because it ensures the
QueryInfo object returned by getQuery always ends up stored in
this.queries.

This divergent behavior was the motivation for calling setQuery in a
number of cases where getQuery would have been sufficient.