Set result.partial only if true in getCurrentResult.
If result.partial is logically false, letting it remain undefined (and
thus falsy) is also allowed by the ?:boolean type of the field. We should
be consistent about omitting the partial field when it's falsy, unless we
want to be consistent about making it always either true or false, which
would likely be a more disruptive change.