Make keyArgs tolerant of optional arguments. (#7109)
As suggested by @tadhglewis in #6973.

The [`keyArgs: ["someArg", "anotherArg"]` configuration for field policies](https://www.apollographql.com/docs/react/caching/cache-field-behavior/#specifying-key-arguments)
should be able to include arguments that are _optional_, while still
specifying an ordering of all possible key arguments for serialization
purposes. When an optional argument is not provided, it will simply
not be included in the serialized `storeFieldName` suffix, and no
exception will be thrown.