Add a paragraph on modifying the response data
I added 2 examples for response manipulation, one that is synchronous and
the other requires async/await. This is important because I could not find
any documentation on how to manipulate the response in an asynchronous
manner, which led me to editing the docs.

See https://github.com/apollographql/apollo-client/issues/7328