Make ApolloLink#onError observer parameter optional. (#7407)
This new parameter could be considered a breaking change (in Apollo
Client v3.3) for existing code that called ApolloLink#onError directly
with only one argument (the error). Fortunately, the implementation of
onError already tolerates the possibility that observer may not be
provided, so we can fix this with a TypeScript-only tweak.